requirements.txt
fastapi
uvicorn[standard]
pillow
pydantic
python-multipart
google-generativeai
python-dotenv
httpx
